Use separate ftp client for listing and downloading cover images
If the ftp client tries to list a directory that it isn't happy with (#recycle seems to really piss it off), it gets into an error state that causes retrbinary calls to fail with a "broken pipe" error code. I didn't see an obvious way to clear this error state, and its unclear what precisely triggers it so that it can be avoided, so for now I am just running retrbinary with a clean instance to bypass the issue.